fix(redaction): avoid UTF-8 boundary panic in diagnostic sanitizer
Closes #2709
