+++ weave.lock
@@ -1,6 +1,6 @@
 [versions]
-csa = "0.1.1107"
-weave = "0.1.1107"
+csa = "0.1.1108"
+weave = "0.1.1108"
 last_migrated_at = "2026-03-08T12:08:01.820964091Z"
 
 [migrations]
